refactor: 패키지 정보 및 수신자 정보 타입 정의를 새로운 파일로 분리하고, 스토어 액션을 리팩토링하여 코드 가독성을 향상시킴
