+++ src/types/recipient-info.ts
@@ -1,0 +1,6 @@
+export interface ReceipientInfo {
+  addressType: 'gs25' | 'cu' | 'custom';
+  address: string;
+  name: string;
+  phone: string;
+}
