+++ src/types/store-types.ts
@@ -1,0 +1,3 @@
+export type StoreType = 'gs25' | 'cu';
+export type ProductType = 'clothes' | 'electronics' | 'books' | 'foods';
+export type WeightType = '500g' | '1000g' | '5000g';
